refactor: UseEnvController hook
